Rewrite API
Label and sequence sinks and sources are now unified. Support for
suspended function is added too. Types are strongly interwoven. Begin
and end of entire elements is now marked.

Support for subtypes is prepared. List and set funnelers are reduced to
the most necessary versions.

Annotation processing is however not implemented, this commit is
preliminary.