Enhance nesting
Nesting may now handle serialization of entire values. The result type
of the sink is now boolean to stop further funneling, the result type of
the source is Nested, where either the actual value, a type substitution
or a continuation flag may be returned. An example is implemented with
serializing to map.