Add after-creation hook
After all arguments of the constructor have been resovled, the item is
created. Then, all variables that are not in the constructor are
assigned. At that point, the item is alread known, a new hook is added
to notify sources that the item they are unfunneled into is now present.